🔨 Refactor: 오늘의 문제 번호 설정 함수 분리
